Retrofit app (OkHttpClient, HttpLoggingInterceptor)

+++ app/src/main/java/com/example/retrofitproject/retrofit/Product.kt
@@ -6,11 +6,10 @@ data class Product(
     val description: String,
     val price: Int,
     val discountPercentage: Float,
-    val rating: Int,
-    val stock: Float,
+    val rating: Float,
+    val stock: Int,
     val brand: String,
     val category: String,
     val thumbnail: String,
     val images: List<String>
-
 )
